--- README.md
@@ -1,2 +1,0 @@
-# App_chinyuki0818
-珍遊記アプリのメインファイルです。（08/18～）
